Rename gem to use use correct convention
From rubygems.org:

> USE UNDERSCORES FOR MULTIPLE WORDS
> If a class or module has multiple words, use underscores to separate
them. This matches the file the user will require, making it easier for
the user to start using your gem.

http://guides.rubygems.org/name-your-gem/

The convention is to use underscores for classes (`ruby_parser` ->
`RubyParser`) and dashes for namespaces (`rdoc-data` -> `RDoc::Data`).

This correction didn't happen in the previous PR (#7) because I was
under the impression that the convention was being broken consistently
in our gems. After auditing half the gems owned by GDS[1] it seems
about 75% are correctly named, so we should keep this up.

[1] https://rubygems.org/profiles/govuk