Don't require active_support
We should not depend on active_support because this app might be used
in projects other than Rails. Instead of accepting both hashes and
symbols, we will assume that the RabbitMQ config given consists only of
keys. It will be the consumers' responsibility to provide the keys.

`@config` isn't used anywhere but in the initializer, so it can be a
local.