Handle invalid JSON
`email-alert-service`, one of the apps that will use this gem, has
special behaviour for when the message received from RMQ isn't valid
JSON. This may be a good thing to pull up into the gem.

We implement the check by refactoring the JSON parsing into a
middleware processor, like `HeartbeatProcessor`. The responsibility of
the `JSONProcessor` is to parse the `payload` attribute on `Message`.

I've considered an alternative, to keep the JSON parsing inside
`Message`, but that would've meant that `Message` would discard itself
whenever it encounters a parse error. Keeping this within a class
seemed cleaner.

`Message` has been refactored to have use `payload` consistently. This
is also the term RabbitMQ uses.