Prevent the ruby consumer from creating rabbitmq queues or bindings
We have moved the responsibility for queue and binding management
to puppet. We might have multiple consumers on a single queue, and
it is then unclear which is the owner of the config.
Also, it has proved easy to misconfigure queue bindings when using the
consumer, since it is not easy to handle binding and queue deletion here.

This removes the routing_key option from the consumer initialiser, and
uses Bunny's no_declare option to avoid creating queues which don't already exist.

See https://trello.com/c/eP2CdnrC/590-fix-rabbitmq-queue-binding

@davidslv and @benhyland