Don't build test files in the gem
There is no need to include test files in a gem that will be used in
production

Whilst it is true that not all gems are production ready, here at GOV.UK
we shouldn't be shipping to production gems that contain test files.

Test files are completely useless in a production environment, besides
provinding them in a production environment can be very dangerous as it
would be very easy to make a mistake.

For example we could be mocking the real Object that is used in
production by accident.

Consider that we have mocked Bunny and accidentally we have executed the
following in production, our RabbitMQ will never receive the data we
were expecting.

```ruby
  queue = instance_double('Bunny::Queue', bind: nil, subscribe: '')
  channel = instance_double('Bunny::Channel', queue: queue, prefetch: nil, topic: nil)
  rabbitmq_connecton = instance_double("Bunny::Session", start: nil, create_channel: channel)

  allow(Bunny).to receive(:new).and_return(rabbitmq_connecton)
```

[[read more](https://github.com/rubygems/rubygems/issues/735)]