Fix batch consumer tests
The tests examples for the BatchConsumer rely on the concurrency
and timings between threads. Tests for the batch consumer's
timeout could intermittently fail. This is due to processor thread
not being scheduled for long enough for the timeout to elapse
before the subsequent messages are sent. This is not an absolute
fix, but should lower the probablity of the test failing.