Reorganise specs to name correct class and idiomatic test location
These tests previously relied on a rather whacky approach of using
`include` on the GovukMessageQueueConsumer module in the spec_helper as
a way to change the class scoping.

That approach itself is a linting violation: `Style/MixinUsage: include
is used at the top level. Use inside class or module.`

It also causes a number of WTFs where in some places classes would be
referenced in their fully qualified form.

This removes that include and then moves the files to a place that
matches where rubocop expects the tests to be. It also updates all the
constants to be explicit references.