Ignore Rspec/FilePath violation
This violation is caused by Rubocop expecting this class to follow a
namespace that matches the class name, which would be the idiomatic way
to organise these files.

Unfortunately the decision was made a while ago to add the test_helpers
directory to lib without a corresponding namespace and changing this
would be a breaking change, which seems unnecessary to resolve for a
linting problem.