Fix broken handling of signals
Our tooling both past (Upstart?) and present (Kubernetes) sends a
`SIGTERM` signal to notify a process that it should gracefully shut
down. At the moment, this library doesn't properly handle this - it does
rescue the error, but does so in the wrong place (so it's not actually
handled) and indiscriminately exits with a non-zero exit code.

- Ensure `SignalException`s raised in `Bunny::Queue#subscribe` (rather
  than the block passed to it) are rescued
- Exit cleanly when receiving a `SignalException` (the comment next to
  `exit(1)` used to say "ensure rabbitmq requeues outstanding messages",
  but that has nothing to do with the exit code, it happens by default
  if the message hasn't been acked by then)
- Only report a signal exception when it's not a vanilla `SIGTERM`
- Add test for signal handling

Ideally in the future we might take this further and allow in-flight
processing to complete before shutting down.