Allow specifying thread pool size and prefetch.
Up to now, we've been forcing workers to be single threaded. This leads
to extremely low CPU utilization for some workloads, for example
search-api-v2-worker, which spends most of its time waiting on the
network for RPCs to Google Vertex.

Bunny is designed for [concurrency] inside consumers.

Allow clients to specify the size of the worker thread pool and the
[prefetch] length (maximum number of unacked messages on the channel).

[concurrency]: https://github.com/ruby-amqp/bunny/blob/-/docs/guides/concurrency.md
[prefetch]: https://www.rabbitmq.com/docs/consumer-prefetch