Add explicit require where `OpenStruct` is used
`OpenStruct` is used by `MockMessage` to represent a message's headers
and delivery information, but the file does not actually require the
`ostruct` library which provides this type.

Many of our projects will somehow transitively require it anyway as it's
extensively used in Ruby, so this omission hasn't caused issues so far,
but we've encountered a problem on `search-api-v2` now where a
completely unrelated dependency was updated and no longer uses
`ostruct`, causing tests that use `MockMessage` to fail as the library
is no longer loaded.